Mensajes de debug; ya funciona parcialmente

+++ hash_crack.py
@@ -1,5 +1,10 @@
 #!/usr/bin/env python3
 
+'''
+TODO: Investigar sobre el uso de Queue o reemplazarlo por flags de true y false:
+
+'''
+
 import hashlib
 import argparse
 import multiprocessing
@@ -13,27 +18,26 @@ def get_arguments():
 
     return parser.parse_args()
 
-def parse_wordlist(wordlist):
-    block = 1000
+def parse_wordlist(wordlist, block_size=1000):
     with open(wordlist, mode="r", encoding="utf-8") as file:
         lines = []
         for line in file:
             lines.append(line)
-            if len(lines) == block:
+            if len(lines) == block_size:
                 yield lines
                 lines = []
         if lines:
             yield lines
-                
+
 def hash_check(lines, algorithm, hash_value, queue):
     for line in lines:
         h = hashlib.new(algorithm)
         h.update(line.strip().encode("utf8"))
+        print(h.hexdigest())
         if h.hexdigest() == hash_value.lower():
-            queue.put(line.strip())
-            break
-        else:
-            queue.put("Debug desde check")
+            queue.put(line)
+            return
+        queue.put(None)
 
 def main():
     args = get_arguments()
@@ -41,29 +45,35 @@ def main():
     processes = []
     threads = int(args.threads) if args.threads else multiprocessing.cpu_count()
     queue = multiprocessing.Queue()
-    
+    #hash_found = False
+
     for block in parse_wordlist(args.wordlist):
-        
-        process = multiprocessing.Process(target=hash_check, args=(block, args.algorithm, args.hash_value,queue))
+        print(block)
+        process = multiprocessing.Process(target=hash_check, args=(block, args.algorithm, args.hash_value, queue))
         process.start()
         processes.append(process)
 
         if len(processes) >= threads:
             for p in processes:
-                p.close()
                 p.join()
             processes = []
-              
-        if not queue.empty():
-            found = queue.get()
-            print(f"Valor: {found}")
-            break
-        else:
-            print("Debug ")
+            
+        value = queue.get() # Si no esta esta linea el script falla
+        while not queue.empty():
+            result = queue.get()
+            print(result)
+            if result:
+                print(f"Valor: {result}")
 
-    for p in processes:
-        p.terminate()
-        p.join()
+                for p in processes:
+                    p.terminate()
+                return
+
+        for p in processes:
+            p.join()
+        
+    #if not hash_found:
+        print("No se encontro el hash en el diccionario")
 
 if __name__ == '__main__':
     main()
